% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_application.R
\name{functions_work}
\alias{functions_work}
\title{Checking whether functions work across different types of objects}
\usage{
functions_work(functions, objects)
}
\arguments{
\item{functions}{blah}

\item{objects}{blah}
}
\description{
Checking whether functions work across different types of objects
}
\examples{
 library(manynet)
 functions_work(c("node_in_optimal", "node_in_betweenness"), 
                c(dplyr::pull(table_data(), dataset))[1:4])
}
